rewrite using astutil.Apply
It's not particularly necessary for the current code, but in the next
commits we want to be able to replace the current node. We can't do that
easily with ast.Inspect.

This API is also more powerful, as it lets us use both pre and post
funcs, and also allows deleting the current node and querying its
parent.