make gofumpt idempotent with imports
It appears we've been running gofumpt in the test scripts as follows:

    gofumpt -w foo.go .

That is, the foo.go file would be reformatted twice. If the tool wasn't
idempotent, the tests might not have been catching that. This was the
case with two test cases in std-imports.go.

The two problems were related to import sorting, and failing to add an
empty line in one edge case. Both are fixed, and the tests now only run
gofumpt once.

Fixes #56.
Fixes #59.