format: rethink goVersion API
Make it have a sane default when given an empty string, so that we don't
have to modify gofmt.go and goimports.go as much. This is important, as
we automatically vendor and modify those files.

This also means that "latest version" is possible without having to
hard-code something like "1.13" or "v99" at the caller site. We handle
this detail internally.