merging adjacent parameters is now behind -extra
Since this rule was more controversial than others; in many cases,
explicitly separate types can be a way to signal that two parameters do
not belong together or are not related.

However, the feature is still useful. For example, when obtaining a
large chunk of generated Go code, or when code is written by new Go
developers who weren't aware of this language feature, it can still be
useful to automatically apply the rule and use tools like 'git diff' or
'git add -p' to manually vet whether or not the changes make sense.

Fixes #53.