format: don't join "C" imports with other imports
The point of the "C" import is to directly follow some commented-out C
code, such as:

	// #include <stdio.h>
	// #include <errno.h>
	import "C"

This can be easily broken by gofumpt if we simply join and sort the "C"
import along with other std imports.

To avoid this, never touch such standalone "C" imports.

Thanks to Manfred Touron for reporting the bug and providing test cases.

Fixes #86.
Closes #87.