format: keep empty lines in funcs if they help readability
For example, the empty line before the body below helps separate the two
lines which belong to different parts of the code but have the exact
same indentation level:

	func multilineParams(p1 string,
		p2 string) {

		println("body")
	}

We could consider enforcing a stricter style for such multi-line
function signatures, such as moving ") {" to the empty line, but it
doesn't seem like it's a universally accepted format.

Moreover, there's nothing really wrong with the format above.

Fixes #108.