leave empty field lists alone if an inline comment is present
If an inline comment is present, joining the lines isn't possible,
anyway. The printer re-adds the newline.

That would be fine, except that it then fails to properly align
comments. This resulted in gofumpt formatting code in a way that's
incompatible with gofmt.

So, just don't remove a newline in this edge case.