format: break long lines if it's easy to do so
Note that this feature is guarded behind GOFUMPT_SPLIT_LONG_LINES=on, as
this feature is still experimental and needs more testing and tweaking
before it's turned on by default.

In summary, if a line takes more than 100 characters, and can be easily
split into two lines of at least about 40 characters, we add a newline.

These numbers are a bit arbitrary and will likely be tweaked over time,
as more testing happens. The basic idea is to split long lines which
have a nice way of being split, but not the others, even if they are
past any line length limit.

There are also some special cases where the rules are bent to result in
better formatting. For the time being, we won't document those, as
they're all still fairly experimental and likely to change.

Updates #2.