don't join declarations with comments
We already skipped joining cases where the comment was in the middle,
such as:

	var foo = ...
	// comment
	var bar = ...

However, we still joined the declarations if it was the first
declaration which had a comment. Thus, the following:

	//go:embed *
	var fs embed.FS
	var s string

would turn into:

	//go:embed *
	var (
		fs embed.FS
		s  string
	)

That's no good, because now the comment directive is misplaced.
This would be bad even if the comment wasn't a directive.

Skip joining any declaration with a comment before it, even if it's the
first one.

The added test case also had an idempotency bug, which is fixed by
running the "multi-line declarations must be separted by an empty line"
logic afterwards.

Fixes #126.