remove codegen to sync with cmd/gofmt
This isn't so much work to do every six months with vimdiff.
Arguably, maintaining this program was more work,
as it broke in subtle ways each release.

Plus, we've started making more aggressive modifications to gofmt,
such as the skipping of vendor directories.
Those get harder to do if the generator needs to be taught about them.