drop -r (rewrite) and -s (simplify) flags
The rewrite flag isn't particularly advanced or powerful.
If some users want it, gofmt will continue to have it.

The simplify flag is on by default in gofumpt,
and we don't really intend the user to turn it off.
Remove the flag entirely, to avoid confusion.

Note that this means we can remove most of rewrite.go,
except the bits needed by simplify.go.

Fixes #105.
Updates #115.