format: always simplify input files
The gofumpt tool has been doing this for a long time,
as we made "gofmt -s" the default behavior.

Unfortunately, since the simplify feature was part of the main package,
direct users of the format APIs were not getting the same behavior.
They were essentially getting the equivalent of "gofumpt -s=false".

Fix this inconsitency by moving the simplify code under format,
and always calling it from the core File API.

Add a test too, which fails if simplification isn't used.

Fixes #115.
Closes #131.