sync gofmt.go with Go 1.18rc1
Mainly to include the work that Bryan Mills and myself did in gofmt
to make it format multiple files concurrently,
resulting in a 3-4x speedup on my 8-core laptop.

Note that this also includes the later change to limit to 200
concurrently open file descriptors, to avoid "too many open files"
errors on some systems.

Also note that we reverted gofmt's transition from interface{} to "any",
as we still want to support Go 1.16 and 1.17.

Update the README to clarify what version of gofmt we're based on.

While here, also update generated.txt to test that gofmt still applies
to implicitly walked generated files, and that stdin is considered an
explicitly given file.