revamp //gofumpt:diagnose for real use cases
Right now, it simply reports "(devel)" as the gofumpt version when
gofumpt is used as a library via the format package,
which is the case for many users like golangci-lint.
To properly test this behavior, we add a gofumpt-external program
which simply formats stdin with one option and prints to stdout.
We `go install` this tool, which depends on a released gofumpt version.

It also does not report the Go version, which is important as we rely on
packages like go/printer, which can influence our behavior.
Similar to the case above, we `go install` a released gofumpt version.

It also does not use VCS information when it is the main module;
add that as well, along with a test, borrowed from mvdan.cc/garble.

While here, make the output easier to read by adding "version" and
"flags" words as descriptions.

Finally, we remove version.Print, which was an unnecessary wrapper.

The following PR will update the released version of gofumpt being used,
meaning that the "released" and "external" tests will pick up the improvement.