use vendored versions of Go 1.19's go/printer and go/doc/comment
These two packages affect the formatting of gofmt, and by extension
gofumpt as well. For example, most changes in formatting to gofmt are
done in go/printer, as it implements most of the formatting logic.

This tight coupling between gofmt and std packages like go/printer
is not a problem for gofmt, as gofmt is released with Go itself.
It is very unlikely that any user would build a version of gofmt with a
different version of those std packages.

gofumpt is a separate Go module, so when users `go install` it, they may
currently be using Go 1.18.x or 1.19.x. This can cause problems, as
choosing a different major version can easily lead to different behavior.

To get the same tight coupling that gofmt has, vendor the two libraries
which affect formatting. go/doc/comment is also in this bag since it
affects how gofmt reformats godoc comments.

Note that this vendoring is not ideal, as it adds code duplication.
This shouldn't bloat the size of gofumpt binaries, as they simply swap
out the original packages for the vendored ones.

For users of the mvdan.cc/gofumpt/format library such as gopls,
they may get duplication if they use packages like go/printer elsewhere.
This is a small amount of unfortunate bloat, but I don't see a better
way to ensure that each version of gofumpt behaves in a consistent way.

Note that we're not vendoring std packages for parsing Go code like
go/parser or go/token, and neither are we vendoring go/ast.
If a project uses new Go syntax, such as generics, they will still need
to build gofumpt with a new enough version of Go.

The copied code comes from Go 1.19.4 and will be updated regularly.
The copying was done manually this time, skipping test files and
rewriting imports as necessary. In the future, we an automate it.
Also note that we ran gofumpt on the copied code.

Fixes #244.