"clothe" naked returns for the sake of clarity
When a function has named results, one can use a "naked return"
to implicitly return those variables. That is, the code

    func Foo() (err error) {
        return
    }

is functionally equivalent to

    func Foo() (err error) {
        return err
    }

Naked returns exist primarily for the sake of avoiding repetition,
but even the Go tour recommends that their use should be limited:

    Naked return statements should be used only in short functions,
    as with the example shown here. They can harm readability in longer functions.

Most Go functions out in the wild aren't short, and in practice,
naked returns tend to be avoided in larger or collaborative codebases
as they confuse the reader about whether any values are being returned,
or which values are being returned if any.

Given that gofumpt is already opinionated, and its use tends to be on
Go modules on the larger side or with multiple contributors,
it seems best to rewrite naked returns to "clothe" them.

Fixes #285.