move "clothe naked returns" under gofumpt -extra
As the discussion in #285 shows, this rule is fairly controversial.
In most cases the change is a strict improvement.
But in some it's a mixed bag. For example, when a func has many results,
some developers use naked returns to avoid long err returns such as:

     return emptyOne{}, emptyTwo{}, emptyThree{}, 0, nil, nil, err

Another point against this rule is that the resulting diffs are noisy,
particularly when compared to other rules ran by default which do
simpler and more localized changes.

The rule is still useful; keep it, but behind -extra.
A set of gofumpt default rules which does 90% of what we want
but can be used everywhere is better than one that does 100%
but is too controversial for half of projects.

Updates #285.